+++ src/mutations/aliasing.ts
@@ -1,4 +1,3 @@
-import * as tsutils from "tsutils";
 import * as ts from "typescript";
 
 import { FileMutationsRequest } from "../mutators/fileMutator";
