+++ src/mutators/builtIn/fixIncompleteTypes/fixIncompleteImplicitGenerics/fixIncompleteImplicitClassGenerics/createExplicitGenericType.ts
@@ -7,6 +7,9 @@ import { FileMutationsRequest } from "../../../../fileMutator";
 import { GenericClassDetails } from "./getGenericClassDetails";
 import { VariableWithImplicitGeneric } from "./implicitGenericTypes";
 
+/**
+ * @returns Test insertion mutation if a generic type should be made explicit.
+ */
 export const createExplicitGenericType = (
     request: FileMutationsRequest,
     node: VariableWithImplicitGeneric,
@@ -15,6 +18,7 @@ export const createExplicitGenericType = (
 ) => {
     const allTypeArgumentTypes: ts.Type[][] = [];
 
+    // For each type parameter, collect the types assigned to it
     for (const typeParameterName of genericClassDetails.typeParameterNames) {
         const typeArgumentTypes = allAssignedGenericTypes.get(typeParameterName);
         if (typeArgumentTypes === undefined) {
@@ -24,6 +28,12 @@ export const createExplicitGenericType = (
         allTypeArgumentTypes.push(typeArgumentTypes);
     }
 
+    // If we couldn't find any types, there's nothing to do here
+    if (allTypeArgumentTypes.length === 0) {
+        return undefined;
+    }
+
+    // Convert the container and its type assignments into a labeled type
     const joinedGenericType = joinIntoGenericType(request, genericClassDetails.containerType, allTypeArgumentTypes);
     if (joinedGenericType === undefined) {
         return undefined;
